remove unused deps, split zokrates_test
