+++ zokrates_test/Cargo.toml
@@ -11,7 +11,5 @@ zokrates_fs_resolver = { version = "0.5", path = "../zokrates_fs_resolver" }
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
-glob = "0.3.0"
-fs_extra = "1.1.0"
 
 [lib]
